correction of the split
